♻️ Extract note properties panel
